Save stat_period_id when saving league games.
This is so that if there are multiple active seasons with the same game type, the game's stats can be applied to only the season's stat period and lifetime/forever period.

Update the ss.refresh_player_versus_stats function to handle refreshing properly when the stat period being refreshed is for a League Season.

Update league.get_season_details to send more detailed game type information, so that it can allow refreshing of player stats (when the game type is for the Team Versus game mode).